Separate periodic task tests
