+++ tests/test_base.py
@@ -9,7 +9,7 @@ import time
 from multiprocessing import Pool, Process
 
 from tasktiger import (JobTimeoutException, StopRetry, Task, TaskNotFound,
-                       Worker, exponential, fixed, linear, periodic)
+                       Worker, exponential, fixed, linear)
 from tasktiger._internal import serialize_func_name
 
 from .config import DELAY
@@ -31,17 +31,12 @@ class BaseTestCase:
         self.conn.flushdb()
 
     def _ensure_queues(self, queued=None, active=None, error=None,
-                       scheduled=None, ignore_queues=[]):
+                       scheduled=None):
 
-        def _ensure_queue(typ, data, ignore_queues):
+        def _ensure_queue(typ, data):
             data = data or {}
-
-            data_names = set(name for name, n in data.items()
-                             if n and name not in ignore_queues)
-            filtered_queues = set(name for name in
-                                  self.conn.smembers('t:%s' % typ)
-                                  if name not in ignore_queues)
-            assert filtered_queues == data_names
+            data_names = set(name for name, n in data.items() if n)
+            assert self.conn.smembers('t:%s' % typ) == data_names
             ret = {}
             for name, n in data.items():
                 task_ids = self.conn.zrange('t:%s:%s' % (typ, name), 0, -1)
@@ -52,10 +47,10 @@ class BaseTestCase:
             return ret
 
         return {
-            'queued': _ensure_queue('queued', queued, ignore_queues),
-            'active': _ensure_queue('active', active, ignore_queues),
-            'error': _ensure_queue('error', error, ignore_queues),
-            'scheduled': _ensure_queue('scheduled', scheduled, ignore_queues),
+            'queued': _ensure_queue('queued', queued),
+            'active': _ensure_queue('active', active),
+            'error': _ensure_queue('error', error),
+            'scheduled': _ensure_queue('scheduled', scheduled),
         }
 
 
@@ -962,101 +957,6 @@ class TestReliability(BaseTestCase):
             assert "not found" in errors[0]
 
 
-class TestPeriodicTasks(BaseTestCase):
-    def test_periodic_schedule(self):
-        """
-        Test the periodic() schedule function.
-        """
-        dt = datetime.datetime(2010, 1, 1)
-
-        f = periodic(seconds=1)
-        assert f[0](dt, *f[1]) == datetime.datetime(2010, 1, 1, 0, 0, 1)
-
-        f = periodic(minutes=1)
-        assert f[0](dt, *f[1]) == datetime.datetime(2010, 1, 1, 0, 1)
-
-        f = periodic(hours=1)
-        assert f[0](dt, *f[1]) == datetime.datetime(2010, 1, 1, 1)
-
-        f = periodic(days=1)
-        assert f[0](dt, *f[1]) == datetime.datetime(2010, 1, 2)
-
-        f = periodic(weeks=1)
-        # 2010-01-02 is a Saturday
-        assert f[0](dt, *f[1]) == datetime.datetime(2010, 1, 2)
-
-        f = periodic(weeks=1, start_date=datetime.datetime(2000, 1, 2))
-        # 2000-01-02 and 2010-01-02 are Sundays
-        assert f[0](dt, *f[1]) == datetime.datetime(2010, 1, 3)
-
-        f = periodic(seconds=1, minutes=2, hours=3, start_date=dt)
-        assert f[0](dt, *f[1]) == datetime.datetime(2010, 1, 1, 3, 2, 1)
-        # Make sure we return the start_date if the current date is earlier.
-        assert f[0](datetime.datetime(1990, 1, 1), *f[1]) == dt
-
-        f = periodic(minutes=1, end_date=dt)
-        assert (f[0](datetime.datetime(2009, 12, 31, 23, 58), *f[1]) ==
-                              datetime.datetime(2009, 12, 31, 23, 59))
-
-        f = periodic(minutes=1, end_date=dt)
-        assert (f[0](datetime.datetime(2009, 12, 31, 23, 59), *f[1]) ==
-                              datetime.datetime(2010,  1,  1,  0,  0))
-
-        f = periodic(minutes=1, end_date=dt)
-        assert f[0](datetime.datetime(2010,  1,  1,  0,  0), *f[1]) == None
-
-        f = periodic(minutes=1, end_date=dt)
-        assert f[0](datetime.datetime(2010,  1,  1,  0,  1), *f[1]) == None
-
-    def test_periodic_execution(self):
-        """
-        Test periodic task execution.
-        """
-
-        # After the first worker run, the periodic task will be queued.
-        # Note that since periodic tasks register with the Tiger instance, it
-        # must be the same instance that was used to decorate the task. We
-        # therefore use `tiger` from the tasks module instead of `self.tiger`.
-        self._ensure_queues()
-        Worker(tiger).run(once=True)
-
-        # NOTE: When the worker is started just before the second elapses,
-        # it's possible that the periodic task is in "queued" state instead
-        # of "scheduled" to ensure immediate execution. We capture this
-        # condition by running the task, and retry.
-        try:
-            self._ensure_queues(scheduled={'periodic': 1})
-        except AssertionError:
-            Worker(tiger).run(once=True)
-            assert int(self.conn.get('period_count')) == 1
-            self.conn.delete('period_count')
-            self._ensure_queues(scheduled={'periodic': 1})
-
-        def ensure_run(n):
-            # Run worker twice (once to move from scheduled to queued, and once
-            # to execute the task)
-            Worker(tiger).run(once=True)
-            self._ensure_queues(queued={'periodic': 1})
-            Worker(tiger).run(once=True)
-            self._ensure_queues(scheduled={'periodic': 1})
-
-            assert int(self.conn.get('period_count')) == n
-
-            # The task is requeued for the next period
-            self._ensure_queues(scheduled={'periodic': 1})
-
-        # Sleep until the next second
-        now = datetime.datetime.utcnow()
-        time.sleep(1-now.microsecond/10.**6)
-
-        ensure_run(1)
-
-        # Within less than a second, the task will be processed again.
-        time.sleep(1)
-
-        ensure_run(2)
-
-
 class TestSingleWorkerQueue(BaseTestCase):
     """Single Worker Queue tests."""
 
@@ -1068,7 +968,7 @@ class TestSingleWorkerQueue(BaseTestCase):
         task.delay()
         task = Task(self.tiger, long_task_ok, queue='swq')
         task.delay()
-        self._ensure_queues(queued={'swq': 2}, ignore_queues=['periodic'])
+        self._ensure_queues(queued={'swq': 2})
 
         # Start a worker and wait until it starts processing.
         # It should start processing one task and hold a lock on the queue
@@ -1078,8 +978,7 @@ class TestSingleWorkerQueue(BaseTestCase):
 
         # This worker should fail to get the queue lock and exit immediately
         Worker(tiger).run(once=True, force_once=True)
-        self._ensure_queues(active={'swq': 1}, queued={'swq': 1},
-                            ignore_queues=['periodic'])
+        self._ensure_queues(active={'swq': 1}, queued={'swq': 1})
         # Wait for external worker
         worker.join()
 
@@ -1089,7 +988,7 @@ class TestSingleWorkerQueue(BaseTestCase):
         task.delay()
         task = Task(self.tiger, long_task_ok, queue='not_swq')
         task.delay()
-        self._ensure_queues(queued={'not_swq': 2}, ignore_queues=['periodic'])
+        self._ensure_queues(queued={'not_swq': 2})
 
         # Start a worker and wait until it starts processing.
         # It should start processing one task
@@ -1101,6 +1000,6 @@ class TestSingleWorkerQueue(BaseTestCase):
         Worker(tiger).run(once=True, force_once=True)
 
         # Queues should be empty
-        self._ensure_queues(ignore_queues=['periodic'])
+        self._ensure_queues()
 
         worker.join()
